+++ pages/about.jsx
@@ -1,6 +1,6 @@
 // aboutページ
 
-import Meta from '@/src/components/meta/Meta'
+import Meta from '@/src/components/elements/meta/Meta'
 import Container from '@/src/components/layouts/container/Container'
 import Profile from '@/src/components/profile/Profile'
 import Hero from '@/src/components/elements/hero/Hero'
