+++ pages/articles/categories/[slug].jsx
@@ -4,7 +4,7 @@ import { getAllCategories, getAllPostByCategory } from '@/libs/api'
 import Container from '@/src/components/layouts/container/Container'
 import PostHeader from '@/src/components/layouts/post-header/PostHeader'
 import Posts from '@/src/components/posts/Posts'
-import Meta from '@/src/components/meta/Meta'
+import Meta from '@/src/components/elements/meta/Meta'
 import CategoriesList from '@/src/components/elements/categoires-list/CategoriesList'
 
 export default function Category({ icon, name, posts, allCategories }) {
